Ajuste de módulo de gerenciamento para usuário comum.

+++ app/Http/Controllers/VehiclesController.php
@@ -4,11 +4,15 @@ namespace App\Http\Controllers;
 
 use App\Driver;
 use App\Setting;
+use App\TrackerType;
+use App\User;
 use App\Vehicle;
 use Illuminate\Http\Request;
 
 class VehiclesController extends Controller
 {
+    private $user;
+    private $final;
     /**
      * Create a new controller instance.
      *
@@ -17,7 +21,11 @@ class VehiclesController extends Controller
     public function __construct()
     {
         $this->middleware('auth');
-        $this->middleware('role:admin|subadmin');
+        $this->middleware(function ($request, $next) {
+            $this->user = \Auth::user();
+            $this->final = !$this->user->isAdmin() && !$this->user->isSubAdmin();
+            return $next($request);
+        });
     }
 
     /**
@@ -27,8 +35,9 @@ class VehiclesController extends Controller
      */
     public function index()
     {
-        $vehicles = Vehicle::where(['created_by' => \Auth::user()->id])->paginate(Setting::paginacao());
-        return view('vehicles.index', ['vehicles' => $vehicles]);
+        $key = $this->final ? 'final_user_id' : 'created_by';
+        $vehicles = Vehicle::where([$key => $this->user->id])->paginate(Setting::paginacao());
+        return view('vehicles.index', ['final'=>$this->final, 'vehicles' => $vehicles]);
     }
 
     /**
@@ -39,7 +48,10 @@ class VehiclesController extends Controller
      */
     public function store(Request $request)
     {
-        if (Vehicle::create($request->input())) {
+        if ($this->final) return redirect(route('vehicles.index'));
+        $all = $request->input();
+        $all["odometer"]= \Helper::stringToFloat($all["odometer"]);
+        if (Vehicle::create($all)) {
             return redirect(route('vehicles.index'));
         } else {
             return $this->create();
@@ -53,8 +65,14 @@ class VehiclesController extends Controller
      */
     public function create()
     {
-        $drivers = Driver::all();
-        return view('vehicles.form', ['vehicle' => new Vehicle(), 'drivers' => $drivers]);
+        if ($this->final) return redirect(route('vehicles.index'));
+        return view('vehicles.form', [
+            'final'=>$this->final,
+            'trackerTypes'=>TrackerType::all(),
+            'finals'=>User::where(["created_by" => $this->user->id])->get(),
+            'vehicle' => new Vehicle(),
+            'drivers' => Driver::all()
+        ]);
     }
 
     /**
@@ -63,8 +81,15 @@ class VehiclesController extends Controller
      * @param  \Vehicle $vehicle
      * @return \Illuminate\Http\Response
      */
-    public function show(Vehicle $vehicle)
+    public function show(Vehicle $_vehicle)
     {
+        return redirect(route('vehicles.index'));
+    }
+
+    private function validateVehicle($vehicle){
+        if ($this->final && $vehicle->final_user_id != $this->user->id) return redirect(route('vehicles.index'));
+        if (!$this->final && $vehicle->created_by != $this->user->id) return redirect(route('vehicles.index'));
+        return false;
     }
 
     /**
@@ -76,11 +101,10 @@ class VehiclesController extends Controller
      */
     public function update(Request $request, Vehicle $vehicle)
     {
-        if ($vehicle->update($request->input())) {
-            return redirect(route('vehicles.index'));
-        } else {
-            return $this->edit($vehicle);
-        }
+        if ($redirect = $this->validateVehicle($vehicle)) return $redirect;
+        $all = $request->input();
+        $all["odometer"] = \Helper::stringToFloat($all["odometer"]);
+        return ($vehicle->update($all)) ? redirect(route('vehicles.index')) : $this->edit($vehicle);
     }
 
     /**
@@ -91,8 +115,14 @@ class VehiclesController extends Controller
      */
     public function edit(Vehicle $vehicle)
     {
-        $drivers = Driver::all();
-        return view('vehicles.form', ['vehicle' => $vehicle, 'drivers' => $drivers]);
+        if ($redirect = $this->validateVehicle($vehicle)) return $redirect;
+        return view('vehicles.form', [
+            'final'=>$this->final,
+            'trackerTypes'=>  TrackerType::all(),
+            'finals'=> User::where(["created_by" => $this->user->id])->get(),
+            'vehicle' => $vehicle,
+            'drivers' => Driver::all()
+        ]);
     }
 
     /**
@@ -103,6 +133,7 @@ class VehiclesController extends Controller
      */
     public function destroy(Vehicle $vehicle)
     {
+        if ($redirect = $this->validateVehicle($vehicle)) return $redirect;
         $vehicle->delete();
         return redirect(route('vehicles.index'));
     }
